fix(fetcher): move tx_fetcher to another class
