+++ lib/bitcoin/fetcher/fetcher.rb
@@ -1,0 +1,5 @@
+class Fetcher
+  def self.fetch(tx_id, testnet: false)
+    raise NotImplementedError
+  end
+end
