+++ spec/bitcoin/tx_spec.rb
@@ -98,26 +98,12 @@ a9143c82d7df364eb6c75be8c80df2b3eda8db57397088ac46430600"
     _bytes.unpack1("H*")
   end
 
-  describe '#TxFetcher' do
-    it 'get base url' do
-      expect(described_class::TxFetcher.base_url).to eq 'https://blockstream.info/api'
-    end
-
-    it 'fetch the wanted tx' do
-      tx_id = '5fbb9471b8b801847274b0e3c51bc76957031293422f5eafe43101d91f2c9a1d'
-
-      tx_fetched_id = described_class::TxFetcher.fetch(tx_id, testnet: true).id.first
-
-      expect(tx_fetched_id).to eq tx_id
-    end
-  end
-
   describe '#sig_hash' do
     let(:tx) { described_class.parse raw_tx, tx_fetcher: tx_fetcher }
     let(:raw_tx) do
-        hex_to_byte_stream(
-          "0100000001813f79011acb80925dfe69b3def355fe914bd1d96a3f5f71bf8303c6a989c7d1000000006b483045022100ed81ff192e75a3fd2304004dcadb746fa5e24c5031ccfcf21320b0277457c98f02207a986d955c6e0cb35d446a89d3f56100f4d7f67801c31967743a9c8e10615bed01210349fc4e631e3624a545de3f89f5d8684c7b8138bd94bdd531d2e213bf016b278afeffffff02a135ef01000000001976a914bc3b654dca7e56b04dca18f2566cdaf02e8d9ada88ac99c39800000000001976a9141c4bc762dd5423e332166702cb75f40df79fea1288ac19430600"
-        )
+      hex_to_byte_stream(
+        "0100000001813f79011acb80925dfe69b3def355fe914bd1d96a3f5f71bf8303c6a989c7d1000000006b483045022100ed81ff192e75a3fd2304004dcadb746fa5e24c5031ccfcf21320b0277457c98f02207a986d955c6e0cb35d446a89d3f56100f4d7f67801c31967743a9c8e10615bed01210349fc4e631e3624a545de3f89f5d8684c7b8138bd94bdd531d2e213bf016b278afeffffff02a135ef01000000001976a914bc3b654dca7e56b04dca18f2566cdaf02e8d9ada88ac99c39800000000001976a9141c4bc762dd5423e332166702cb75f40df79fea1288ac19430600"
+      )
     end
 
     it 'returns the correct sig_hash' do
